homework4/requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
tqdm
segmentation_models_pytorch
opencv-python